[workspace-refactor] Centralize utilities in dagster._cli.utils
